Make use of C-style complex numbers consistent: use 'double_complex' everywhere, which is a typedef for 'double complex'.
Previously the old code used 'double_complex', 'double complex' and 'complex double' interchangeably (the last one is a GCC extension!!).
We also need 'double_complex' to work in C++ code because the LFC-GPU code needs it (via LFCObject).
We handle this by defining 'double_complex' as an alias to std::complex<double> if compiling as C++. This is OK because these complex types have same identical memory layout.